requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
requests
APScheduler
reportlab
pytz